added secret-init containter & refactored server
