+++ cmd/commands/server_secret_init.go
@@ -1,0 +1,36 @@
+/*
+Copyright 2022 The Numaproj Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package commands
+
+import (
+	"github.com/spf13/cobra"
+
+	secretinitcmd "github.com/numaproj/numaflow/server/cmd/serversecretinit"
+)
+
+func NewServerSecretInitCommand() *cobra.Command {
+
+	command := &cobra.Command{
+		Use:   "secret-init",
+		Short: "Initialize secret with admin password and jwt secret key",
+		Run: func(cmd *cobra.Command, args []string) {
+			secretinitcmd.Start()
+		},
+	}
+
+	return command
+}
